requirements.txt
numpy
torch
torchvision
tqdm
matplotlib
colorama
pillow
argparse
git+https://github.com/openai/CLIP.git